Return after callback is moved and run in |DownloadImageFromAxNode|
This issue was identified by a code scanner at Microsoft. There is an
early exit condition in |ImageDownloaderImpl::DownloadImageFromAxNode|
that moves and run the function callback but doesn't actually return
from the function. This CL adds a return after the callback is run for
this condition.

Bug: 392042079
Change-Id: I2249aff631fc6e1273e68d7bac72fcce4863f63a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6196280
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Jacob Francis <francisjp@google.com>
Commit-Queue: Sanket Joshi <sajos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1411161}
NOKEYCHECK=True
GitOrigin-RevId: 6e105f03ee46d2d6ea41cfc1ed1874c57cdd0e67