Make hit testing for a flex block with only one child element correctly
This patch[1] changes `ShouldUsePositionForPointInBlockFlowDirection`.
If it is not a layout block flow, then return false. There is a
special case, consider this test case:
<div id="outer" style="display: flex;">
  <div contenteditable>Hello |World</div>
</div>
When clicking outside the div#outer, it was expected that the caret
would disappear and the div#outer would blur. However, the caret
actually move to `World|`. The hit testing result of the click event
is `Hello World#[11]`, because the div#outer is a flex block not
a layout block flow.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/3067117

Bug: 40889098
Change-Id: I72cf5fcef7c2bf44d186244dd1c735656ab51ced
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6196657
Commit-Queue: Peng Zhou <zhoupeng.1996@bytedance.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411496}
NOKEYCHECK=True
GitOrigin-RevId: 7652b5e315c098244e401b6cbdb560962d9bf5ec