Split out universal selectors as a separate match type.
Currently, we treat the universal selector * as a tag match;
however, it is distinct from tag matches in a variety of ways
(e.g., it has different specificity). We split this out into
a new match type kUniversalTag (it's still called “tag” because
the namespace part, if any, will match on the tag name).
This also makes it much easier for easy selector matching
to support it, allowing selectors like `[dir="ltr" i]` from
the UA sheet (which has an implicit universal selector for
matching the HTML namespace) to be treated as easy selectors.

Both style perftest and Speedometer3 is either neutral or
slightly positive.

Change-Id: If0070b93da08b8931fbd7d7425a831c1a5fb66e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6197896
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411514}
NOKEYCHECK=True
GitOrigin-RevId: 28b4b79eac9b7cf46bbe559b59f8a40c1bc1f39e