[Blink] Add CanvasRenderingContextHost getter for SharedImageFormat
Will be used to convert callsites of
CRCHost::GetRenderingContextSkColorType() to instead operate on the
SharedImageFormat. A bit down the line, we will swap the layering so
that the SkColorType getter is on top of the SharedImageFormat getter
(and then hopefully goes away entirely).

Bug: 371227617
Change-Id: Ief61768c671531d6d8c67a21f31edbe09e5b5afa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6198502
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411570}
NOKEYCHECK=True
GitOrigin-RevId: 0006ebf3ef8a1969f5eacb02f669cb1513ee7bdf