Precompute most of MatchRequest.
Compute the set of RuleSets (we call it a RuleSetGroup) that we need
for a given set of stylesheets (typically “all author stylesheets”,
“all user stylesheets” or “a given combination of UA stylesheets”)
ahead of time instead of on-the-fly. We move it to Oilpan in the process,
since it's now permanent.

This both saves a bit of time (especially clearing the array of
RuleSet pointers during initialization is a pain), and allows us to
do more expensive precomputation in the coming patches. It means
we need to iterate over bitmaps instead of just integers, though,
which is slightly more expensive. (There are some cases, like
RuleSet-based invalidation, where building the MatchRequest cannot
be cached and the initialization then is somewhat more expensive
than it used to be.)

This patch is just the base functionality, keeping most things equal
to before; the only thing we really save on is that we don't have
to iterate over all RuleSets every time to see if they are
single-@scope or not (we store that in a bitmap). The initial
performance numbers will come in the next patch, which has been
split out for easier revertibility and reviewability.

In order to increase confidence in the correctness, we add DCHECKs
that the cache delivers the same RuleSetGroups as without the cache.
We've run through 10k top pages without this DCHECK hitting.

Change-Id: Ia578f267ec082798af9e2937e9d1b88939abedaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179411
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411600}
NOKEYCHECK=True
GitOrigin-RevId: 19acd4c22682c869301d31c12f545c85088d1ec6