Rebaseline pinch-zoom web test
The test is failing on mac due to a small patch of pixels that is one
shade of gray different. This CL rebaselines the image to match the
expectations.
It also increases the fuzziness tolerance to let the test pass if this
problem crops up again.

Bug: 392524783
Change-Id: I75d7c059db430f81f47cdbeaa309f011694af9e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6201832
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1411723}
NOKEYCHECK=True
GitOrigin-RevId: 1503e6e4e09a4bd545668eb19b282219f78f75a4