[A11y] Allow children for br elements in all cases
Line breaks in the accessibility tree normally contain a single inline
textbox child. However, if a script adds a DOM child to the line break,
that child is not added to the accessibility tree, and the textbox child
is removed.

This removal causes a mismatch between the accessibility tree's node
count (from the AXObjectCache) and the serialized tree's node count
(from the AXTreeSerializer).

Since Blink doesn't render script-added DOM children on line breaks, and
these children aren't in the layout tree, always including the inline
textbox child in the accessibility tree (even when a script adds a DOM
child) resolves the node count mismatch and aligns the accessibility
tree more closely with the layout tree.

Bug: 382551990,382732986
Change-Id: I65d6e88ea3489096948c726dcfc582bf2f5ca33c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6108264
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Jocelyn Tran <jocelyntran@google.com>
Cr-Commit-Position: refs/heads/main@{#1411753}
NOKEYCHECK=True
GitOrigin-RevId: 8755f33cb11a1a1d041c3e0687b2a4524a119627