Add :has-interest pseudo class for interesttarget feature [9/N]
This allows styling the element with the interesttarget attribute
depending on whether it currently has interest.

Also, one of the desired features is to easily be able to change the
show/hide delays for interesttarget depending on whether another
element already has interest. For example, when multiple elements
have interesttarget popovers, once the first one is showing, it
is common for the next one to show up more quickly or even
instantly. With the new pseudo class, this is possible via:

  [interesttarget] {
    interest-target-delay: 1s;
  }
  container:has(:has-interest) [interesttarget] {
    interest-target-show-delay: 0s;
  }

Bug: 326681249
Change-Id: I98d34e0f1a06bb6d80b153f3839678ada97aac08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6199201
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411801}
NOKEYCHECK=True
GitOrigin-RevId: 9d53f647a39d0965480e0ea16dd22002b7a5e3ce