Blink: Have canvas test not rely on GetImage creating resource provider
A CanvasRenderingContext2D unittest currently invokes
CanvasRenderingContext2D::GetImage() to trigger creation of the
CanvasResourceProvider. However, we will shortly be changing this method
under a killswitch to return nullptr if the CanvasResourceProvider isn't
present in order to align with behavior of
OffscreenCanvasRenderingContext2D and expected semantics of the method
that it returns nullptr if "the buffer backing the canvas" isn't
present.

This CL changes this test to invoke `DrawSomething()` to create the
resource provider, as other tests here do. Note that the current usage
of GetImage() to create the resource provider is incidental, i.e. the
test is *not* testing that GetImage() causes the resource provider to
be created, it just needs the resource provider to be created as part
of the setup for what the test is testing.

Bug: 40280152
Change-Id: Ia70aa4a35b8041f591f8f0477f8744e67287cbb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6198768
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1412135}
NOKEYCHECK=True
GitOrigin-RevId: 9da67427c03c138fb81149dc86fffd2cadf145c8