[line-clamp] Test that the ellipsis behaves like `line-height: 0`
The spec text for `block-ellipsis`, which describes the presence and
behavior of the line-clamp ellipsis, says that that ellipsis is
wrapped in an anonymous inline which is a direct child of the root
inline box, and which is assigned `unicode-bidi: plaintext` and
`line-height: 0`.

We don't implement the ellipsis as an inline, let alone one with
computed styles. However, the `line-height: 0` requirement means that
the ellipsis should not cause the line box's height to grow, which is
already the case in our implementation. However, this was not tested
in WPT (or in Chrome-specific tests) until now.

Since the ellipsis inline would have the same font and font-size
styles as the IFC root, and it only contains text, the only way that
it could possibly grow the line box's height would be if the text uses
a fallback font with taller metrics. Since Gentium Plus happens to
have taller font metrics than Ahem, we use these fonts in `@font-face`
declarations with the `unicode-range` property to make sure the
ellipsis renders with Gentium Plus rather than Ahem.

Bug: 40336192
Change-Id: I2ba375cd25cde8444a848243bbe287ddc7cdafba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6198346
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1412186}
NOKEYCHECK=True
GitOrigin-RevId: fa359c415b793fc65635d58b4b08eee323307946