[functions] Refactor ResolveVarInto to handle fallback first
As a preparation for handling fallbacks for arguments and locals,
resolve the fallback part of a var() before doing the main lookup.
This makes the `fallback` TokenSequence already available when
arguments and locals are handled, which a future CL will utilize.

The performance characteristics of the function should not change
substantially, since we already resolved the fallback unconditionally
(to look for cycles).

There should be no behavior change.

Bug: 325504770
Change-Id: I739c26519b6a306d445718df28e580c288b11d0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6187537
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1412199}
NOKEYCHECK=True
GitOrigin-RevId: 0f7f8ad5a20ac582b5a0279c2302b5ef57e48e8a