Invalidation tracing: Add parent and pseudo IDs
This is a follow-up to https://crrev.com/c/6176521 to help account for
every node affected during a Recalculate Style event. Since subtree
invalidations are only marked on the root of the affected subtree, we
need a way to follow descendant recalculations up to the marked node.
During live analysis we could simply traverse the tree, but in a saved
trace we don't have that option.

Additionally, a given element ID might be logged more than once during a
given Recalculate Style event if pseudos attached to it are also
recalculated.

To help with both of these cases, this CL adds parent and pseudo ID
information to the existing ResolveStyle event.

Bug: 337076014
Change-Id: Iabb241a4a599171217b1b95332d841aa7f60766b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6173765
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1412416}
NOKEYCHECK=True
GitOrigin-RevId: 887bb5c009e6d452dbc20b6a3f6b01dd7d42dcae