Check for content descendants in <selectedcontent>
In this CL, a check is added to `TraverseAncestorsAndCheckDescendant` to
account for the case of `<selectedcontent>` being an ancestor of cloned
content from an `<option>`. Previously, content of `<selectedcontent>`
fired an accessibility console error, so a test was added to validate
that this is no longer the case.

Bug: 390477775
Change-Id: I53e838e3488f77a3d42217ddaf6144cfe643da9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6203671
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Ana Sollano Kim <ansollan@microsoft.com>
Reviewed-by: Sanket Joshi <sajos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1412457}
NOKEYCHECK=True
GitOrigin-RevId: 0b0907c5349ad36c63adc7c1e5d045adee5366fa