[shape-cache] Make `FontFeatures::Initial()` thread-safe
Because this code is expected to be called from shaping,
which could be in a background thread, this patch makes it
thread-safe.

Also it makes `IsInitial()` not to depend on `Initial()` for a
minor performance improvement.

Bug: 377832797
Change-Id: If25413397be9b3732170798a2639f979ebbc0f65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6208828
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1412498}
NOKEYCHECK=True
GitOrigin-RevId: e28d63c2514fbd78af3b22d76763efb70c17b67f