Fix customizable select slot assignment recalc crashes
Now that we don't change slotting based on whether the popover is
showing or not, we can remove the calls to make slot assignment dirty
which were causing a slot assignment forbidden crash in accessibility
code.

This patch also adds a double rAF to the switch-picker-appearance test
to make it wait for the picker to get closed after the task gets posted
in PopoverelementForAppearanceBase::DidRecalcStyle.

Fixed: 388115419, 384394713
Change-Id: Iaf3a488597b19a7f90d7977df24e74ef5f433996
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6166184
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1412579}
NOKEYCHECK=True
GitOrigin-RevId: 51bcbf7d7dfc806341623efafdb75bfcfffbd064