Improve local-floor handling
Adds a mojo_from_floor transform and appropriately uses it in blink.
Prior to this, local-floor used `VRStageParameters` to get the floor
height, even on platforms like ARCore that don't support a stage. The
stage parameters and estimated height are used as a last resort fallback
in blink, since the WebXR spec requires that the local-floor space must
be supported, even if it is emulated in this way.

ArCore is updated to simply send mojo_from_floor instead of a full set
of stage parameters.

OpenXR also now promotes local-floor to a first-class XrSpace. This is
done either via the LOCAL_FLOOR_EXT extension if it is present, or by
creating an offset space of either local_from_stage or a fixed height
offset. This allowed simplification of some anchor handling code, and
improves the local-floor experience on devices where it is supported.

Bug: 390216623
Change-Id: If32a46a78e13955bdcfab3b332c2ad754c7564fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6182257
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1412583}
NOKEYCHECK=True
GitOrigin-RevId: 29e085ea975f805b67fdbec7ec51dc524ce484ea