Remove `HangingWhitespaceDoesNotDependOnAlignment` feature flag
This flag was added in https://crrev.com/c/4881393 and shipped in
Chromium 119, to roll out the change if there was any breakage.
Although that CL was based on a misunderstanding of the spec, which
was fixed in https://crrev.com/c/5147580 (Chromium 124), that fix was
still guarded under the same flag because there might still have been
breakage. At this point, there has not been any breakage from this
second patch, and the feature flag can be removed.

Alongside the removal of the flag, this patch removes the
`LineInfo::ShouldHangTrailingSpaces()` method, since it would now
always return true, and removes the second argument to the
`AdjustLineOffsetForHanging` function in `inline_layout_algorithm.cc`,
since it would now be unused.

Bug: 1363901, 40944859
Change-Id: Iea3d406ab8ff9cf6a28d3372624a06bcc2b7dc8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6190628
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1412801}
NOKEYCHECK=True
GitOrigin-RevId: 697339ea16c1466671e430b951a4998c51887681