Remove no longer needed anonymous namespace from header file
This file has a struct in the header that's inside an anonymous
namespace which is not allowed in the [style
guide](https://google.github.io/styleguide/cppguide.html#Internal_Linkage).

The original reason for the anonymous namespace was to allow the
DocumentLoader class to declare the struct as a friend. The actual
implementation doesn't require the struct to be a friend of
DocumentLoader, because it doesn't rely on access to private or
protected members or methods in DocumentLoader. Allowing the anonymous
namespace and friend declaration to be removed.

Bug: 392860311
Change-Id: I145146e6502822456c50c3112327c96ec7a6a7ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6207868
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Aaron Selya <selya@google.com>
Cr-Commit-Position: refs/heads/main@{#1412914}
NOKEYCHECK=True
GitOrigin-RevId: 11fb2225781709095704cd0092d41e321e44059a