Fix ListedElement is focusable pre-check for host with delegatesFocus
The function calls are as follow:
HTMLFormElement::reportValidity()
  HTMLFormElement::ValidateInteractively()
    ListedElement::ValidationAnchorOrHostIsFocusable()

This means a custom element inside a form element will call the above
functions to check validity of the form input and show the validation
message accordingly.

In ListedElement::ValidationAnchorOrHostIsFocusable, the function checks
whether the anchor element or the host is focusable. If the anchor is
specified, it checks the correct element. If not, it will check the
host. However, in the case of a custom element with delegatesFocus,
we should not check the host, but instead check its focusable area.

Change-Id: Ie331a17c3622ac949ce70cfcacbf035efd73ae89
Fixed: 389587444
Bug: 40726105
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6174381
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1413006}
NOKEYCHECK=True
GitOrigin-RevId: 77964e11aa7042fdd1341ad5a681ea10abaed5dc