[blink] Remove cancelled scheduler idle tasks when many are accumulated.
With this CL, ScriptedIdleTaskController counts its outstanding
scheduler idle tasks. When many more than the number of
"non-cancelled" IdleTasks are accumulated, it invokes a new
method on ThreadScheduler which traverses the queues of idle tasks
and remove those that are cancelled. The impact is that this code does
not cause an Out of Memory (OOM) crash anymore:

  let noop = () => {}

  let postAndCancel = () => {
   let id = window.requestIdleCallback(noop);
    window.cancelIdleCallback(id);
    scheduler.postTask(postAndCancel);
  }

  postAndCancel();

Bug: 365114039
Change-Id: Id44b773bb54a2bd70a467ebef54142f6dc7c5b42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6205406
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1413248}
NOKEYCHECK=True
GitOrigin-RevId: 90c3b90d0d0f5983d7a80bee74aa835656df1340