Add a Bloom filter for which attributes an element has.
This saves us to going down to the ElementData and reading all
the Attributes stored therein in the fairly common case that
the attribute we're looking for isn't there. (This is especially
interesting for querySelector(); selector matching as part of
regular style usually has this already covered by means of
bucketing attribute rules.) The memory reference can be quite
expensive when traversing a large DOM, and tends to blow up
the L2 cache.

This doesn't increase the size of Element, as there was already
32 padding bits at the end.

The style perftest is mostly neutral, but Speedometer3 shows
clear gains (M1 Pinpoint, LTO but no PGO, significant results
at 99% CI only, lower is better except for score):

  TodoMVC-WebComponents                       [ +0.8%,  +1.1%]
  TodoMVC-JavaScript-ES5                      [ +0.2%,  +2.6%]
  TodoMVC-JavaScript-ES6-Webpack-Complex-DOM  [ -8.4%,  -8.0%]

  Score                                       [ +0.1%,  +0.4%]

It's possible to also include the values in the filter, but it
doesn't seem to gain anything, and would require updating it
much more often.

Change-Id: I27ab08c5ba67f9df8dd1a7d32361ac5ed1475f57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6197390
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1413368}
NOKEYCHECK=True
GitOrigin-RevId: 2463227eadb35cb5a15c5c6860d3a1ffa59a66d2