[SyntheticResponse] Fake service worker registration
This CL adds the fake service worker database that is returned if
SyntheticResponse is enabled, and the request URL is matched to the url
in the allowlist provided from the feature param.

The fake service worker is basically do nothing, but this is necessary
to handle the navigation with the service worker classes in the browser
process e.g. ServiceWorkerMainResourceLoader.

Since SyntheticResponse will be implemented on the service worker infra,
we need this fake service worker registration. After the experiment we
have a plan to remove this from the code base and give developers a way
to control on their side via service workers.

Fuchsia-Binary-Size: Increase is temporary.
Bug: 352578800
Change-Id: I491735ed6ebdf8bb7d357f1f779d5d964ddbb1e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6093911
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1413493}
NOKEYCHECK=True
GitOrigin-RevId: faf40c02d02090a5327eaeee61890b7503f32fe6