Not inherit backface-visibility across 3d rendering contexts
Specifically, an element creating a new 3d rendering context should
not inherit parent backface-visibility.

I'm not sure if this is correct in the spec's perspective, but it
avoids a failure of DCHECK in cc which requires backface inheritance
not to cross 3d rendering contexts.

dbaron@chromium.org added a comment at
https://github.com/w3c/csswg-drafts/issues/918#issuecomment-2625306346
so that we can think about it more in the future if we need to.

Bug: 391504029
Change-Id: I1c4e2b45d0adb0dd116d96373114fde8a53fdb90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6207584
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1413640}
NOKEYCHECK=True
GitOrigin-RevId: 64aa9f099ff1c810b0c61abc6c75676b22e9688a