bindings: Slimmer ScriptValue
ScriptValue is a primitive that is used ubiquitously all over the
place. The Isolate pointer in there was barely used and we can also
store the Isolate in the DOM world which is a much more coarse grained
concept.

S3 seems noisy on this change but it seems generally viable to do this.

Summary:
👍 TodoMVC-JavaScript-ES5                      [ -0.5%,  -0.1%]
👍 TodoMVC-JavaScript-ES6-Webpack-Complex-DOM  [ -0.6%,  -0.1%]
👍 TodoMVC-Vue                                 [ -0.5%,  -0.1%]
👍 TodoMVC-WebComponents                       [ -0.4%,  -0.0%]

Change-Id: I5d4f8780c0b4cf3b288bb4509c79d53660df64c2
Fixed: 393153134
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6212565
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1413657}
NOKEYCHECK=True
GitOrigin-RevId: 7bdbad99dc7518371190ae2f0f21483a7e97aec9