Change order of [inert] vs :modal UA style for interactivity
A fullscreen'ed element or modal dialog should still be inert if it has
an inert attribute. This keeps the existing behavior when the inert
attribute set html inertness.

The specificity is modified to ensure this behavior.

Additionally add an interactivity declaration for select buttons to
prepare for no longer setting html inertness for IsInertRoot().

The UA rule could have been implemented as :modal for both dialogs and
fullscreen elements, but that would have ended up in the universal
bucket for all elements for the dialog case.

Bug: 370065759
Change-Id: I5be974534abb6714cc964070cbf04437408a2a08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6216527
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1413833}
NOKEYCHECK=True
GitOrigin-RevId: 4670338ccf46aaf26aee2402b0f722106d55b003