Remove canonicalization of selection_to_delete in MoveParagraphs()
There are bugs when deleting at the end of a block boundary if there are
non-editable nodes in the second block. When moving the second block,
|MoveParagraph()| function creates selection_to_delete as visible
selection from start and end positions. During this canonicalization,
the selection is adjusted to avoid shadow roots, editing boundaries etc.
This leaves some nodes from being deleted from second block which is not
expected.

Since the start and end are passed as visible positions and they are in
same enclosing block in this context, we can skip the canonicalization
of selection_to_delete.

This canonicalization is not present in
|CompositeEditCommand::MoveParagraphWithClones()|

Bug: 41474791
Change-Id: I60010cb79f2797960684f1617220fff13d27cb88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6214650
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Sambamurthy Bandaru <sambamurthy.bandaru@microsoft.com>
Reviewed-by: Siye Liu <siliu@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1413880}
NOKEYCHECK=True
GitOrigin-RevId: 49f2e7b91890ac5e9bc6c80fa18d0c6a0980aa31