[DC] Support new DC API request format
The new format for DC API requests, has the following renames:
providers --> requests
request --> data

The new format is supported in together with the legacy format with
priority given to the new format when both are available in the request.

In addition, a run time flags is introduced to stop supporting the
legacy format in the future.

Details are in the linked bug.

Bug: 370982995
Change-Id: I2239e2daac06c71fbc7b787c8f2c3d2ee36cee53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5937839
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1414272}
NOKEYCHECK=True
GitOrigin-RevId: 4a0ba6027601b356ed898c997bd4dfcf0b51fd6a