Don't clear ForceReattachLayoutTree state on <slot> with display lock.
This fixes a regression that started occurring frequently with text
between MathJax elements inside of a <details> element (which uses
content-visibility, and thus display lock).  The regression resulted
from changes to the internals of the <details> element.

When detaching the layout tree for this particular case of <slot>
elements with display lock, this restores the ForceReattachLayoutTree
state that existed before so that it's still set when the display lock
is no longer present.

Fixed: 383593255
Change-Id: Ib1dab4da0b1211ecf35dd6e11201cc35064e466a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6220858
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1414486}
NOKEYCHECK=True
GitOrigin-RevId: 9227cdf8a138a0cdc3b3ca9f6deae79c71962ed5