Consider for ImplicitAnchorElement() for ShouldStoreOldStyle()
Elements with multiple recalc passes need to store the old Element
ComputedStyle at the start of the lifecycle update to have the correct
before-change style for transitions. Anchored elements have multiple
recalc passes because of layout based anchor evaluations.
ShouldStoreOldStyle() checked for this, but did not include checking
ImplicitAnchorElement().

In the specific case it meant we ended up with an intermediate null
base style for a display:block -> display:none transition which caused
a crash.

Bug: 392042073
Change-Id: I1718c65b7c91a4a67d611333afd1214313baae22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6219263
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1414784}
NOKEYCHECK=True
GitOrigin-RevId: 0cdc1fda5834112bee26e00847444a99055aa4df