[anchor] Avoid negative IMCB correctly with scroll shift.
When a positioned box is tethered both to an anchor and the original
containing block (e.g. position-area:top), and the relevant anchor edge
is scrolled out of view, the inset-modified containing block size needs
to be clamped to zero. At the same time, the inset that's auto should be
adjusted so that the border box of the positioned element steers clear
of the relevant anchor edges.

We used to check if the inset at the opposite side of the anchor was
zero, and adjust it accordingly if that was the case. But the comparison
with zero wasn't correct when there was a default scroll shift. Rather
than fixing it by comparing to the default scroll shift, clean up the
code to check if the relevant inset is specified or not, which is what
the code really wanted to do.

Bug: 373874012
Change-Id: Ic98f63d8098eb0c8a7192b3bc0d1b5efa5194f74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6218569
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1414807}
NOKEYCHECK=True
GitOrigin-RevId: 57f539e2ec714bd39a0c8b90dd5221762cf40806