Reland "Add a Bloom filter for which attributes an element has."
This reverts commit 49649d88162167c585f28aca695f451f1e0b3243.

The failing tests pass locally (and did pass in SQ, but not in CI).
However, it seems that all of failing tests are about lacking the
synchronization of the style attribute, so we do that now before we
check the filter (i.e., we move the Bloom filter test below the calls
to synchronize the attribute in all cases).

Change-Id: I57ef1223fdf89ff326cfb3fed5a51d1297a0d8f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6218391
Auto-Submit: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1414812}
NOKEYCHECK=True
GitOrigin-RevId: 4c4f9f4a36a42399088b2fc43a450bca74cecbef