Add specific console messages for each issue reason
This CL modifies the console error message to provide more information
on what the error is about and how to fix it. This provides context to
web authors on how to avoid accessibility errors if the content model of
the customizable select isn't followed. Tests are added and expectations
have been updated to verify behavior.

Bug: 354032791
Change-Id: Iea08be97dbf5d1780a0040cc99b1b155a6484909
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6219716
Reviewed-by: Sanket Joshi <sajos@microsoft.com>
Commit-Queue: Ana Sollano Kim <ansollan@microsoft.com>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1415176}
NOKEYCHECK=True
GitOrigin-RevId: 4e5d97b7ca2f95294361a6bf9e4e48273c044295