WebUI: Make CrLitElement#shadowRoot non-nullable.
This allows client code to not have to use a type cast like
`this.shadowRoot!` whenever accessing the ShadowDOM. All CrLitElement
instances are expected to have a populated `shadowRoot` property.

Changing only a couple files, one file in Downloads and a few other that
were throwing TypeScript errors with this change, and will do a bulk
removal of the extraneous "!" in follow-up CLs.

Bug: 393525665
Change-Id: Ic6ae85617eedcc6e3c60c1736dd02f3b24b48aa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6220860
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1415201}
NOKEYCHECK=True
GitOrigin-RevId: dad8505a500bc73c8bcb002860b0964472707659