[SVG] Fix viewport calculation for uninstanced symbol elements.
The test svg-baseval-in-display-none.html verifies the behavior of
the `baseVal` property for SVG elements, including the symbol
element. It ensures that the `baseVal` properties of geometry
elements inside the symbol element are correctly computed.

To compute the correct `baseVal` properties, the viewport size is
required when width and height are given in percentage format.
However, in Chromium, the symbol element (as a viewport element)
always returns an empty viewport size when computing baseVal
properties. This results in the baseVal properties returning zero
values.

The spec[1] says either the `svg` element or a `symbol` element
that is instanced by a `use` element establishes a new viewport.
In the test, the `symbol` isn't instanced, so the expectation should
be that the nearest ancestor `svg` provides the viewport.
In this CL, we compute the viewport size for the `symbol` element
provided by its nearest svg ancestor.

[1] https://svgwg.org/svg2-draft/coords.html#EstablishingANewSVGViewport

Bug: 41463219,41461114
Change-Id: Iac60cb45f1284db78c36a60c54e9dd9c341172ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6217852
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1415367}
NOKEYCHECK=True
GitOrigin-RevId: dcb9f111a2af4f64e1f7361bc3ae1d462de82722