Use IDL enums for all CanvasTextDrawingStyles in Canvas2D
Some of the canvas properties that are part of CanvasTextDrawingStyles
(https://html.spec.whatwg.org/#canvastextdrawingstyles) had been
modified to use enums defined from the IDL definition of the rendering
context (fontStretch and textRendering).

This CL creates enums for the rest of the CanvasTextDrawingStyles
properties. Getters and setters had to be modified, as well as a unit
test. Generated tests for setting and getting this properties already
exist in WPT, as defined in text.yaml.

Bug: 341213359
Change-Id: Ic8c30219a039c96cfd83dd078157d9fb60565e03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6000733
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1415577}
NOKEYCHECK=True
GitOrigin-RevId: 66ceeef1ee9be8a1da3899f6d7df6045689fbe2a