Ignore ancestor clips for scrollers when expanding cull rect
This can be either an alternative of crrev.com/c/6215670, or a
follow-up of it to reduce checkerboarding under-reporting in rare
cases. This can also simplify the logic regardless of the bug.

The code change is behind a killswitch
ScrollCullRectFromContainerRect.

Now if a scrollport is visible in the input cull rect, we always use
the full scrollport before expanding the cull rect for scroll, to
avoid the effect of ancestor clips. This avoids the cause of
crbug.com/391919458 (though fixed by crrev.com/c/6215670), and
simplify logic dealing with ancestor clips (e.g. the logic added in
crrev.com/c/6133753 now not needed).

To avoid too much expansion for scrollers whose only a small part is
in the input cull rect (rare), the small_scroller optimization is
changed to check for the area of the intersection of scrollport with
the input cull rect instead of just the scrollport.

Bug: 391919458
Change-Id: Iae486aa2e3a7f9aa261efa4d376dd4d890006544
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6216280
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1415703}
NOKEYCHECK=True
GitOrigin-RevId: 7acc815c4bfacad6978d799f41610d0c0a1fd5d0